+++ src/just_makeit/_render.py
@@ -669,7 +669,7 @@ MODULE_EXT_C_HEADER = """\
 #include <Python.h>
 #define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 #include <numpy/arrayobject.h>
-#include <complex.h>
+#include "clib_common.h"
 <<module_extra_includes>>
 <<module_core_include>>"""